refactor question translations
